src/modules/orders/services/ShowOrderService.ts
import AppError from '@shared/errors/AppError';
import { injectable, inject } from 'tsyringe';
import Order from '../infra/typeorm/entities/Order';

import IOrdersRepository from '../repositories/IOrdersRepository';

interface IRequest {
  order_id: string;
}

@injectable()
export default class ShowOrderService {
  constructor(
    @inject('OrdersRepository')
    private ordersRepository: IOrdersRepository,
  ) {}

  public async execute({ order_id }: IRequest): Promise<Order> {
    const order = await this.ordersRepository.findById(order_id);
    if (!order) {
      throw new AppError('Order not found', 404);
    }
    return order;
  }
}
